// src/commons/enums/FieldTypeEnum.ts
export enum FieldTypeEnum {
    inputText = 'Input',
    textarea = 'Textarea',
    number = 'InputNumber',
    select = 'Select',
    treeSelect = 'TreeSelect',
    datePicker = 'DatePicker',
    dateRange = 'DateRange',
    switch = 'Switch',
    radio = 'Radio',
    checkbox = 'Checkbox',
    file = 'File',
    staticText = 'StaticText',
}